scan ticket at an attraction => increases queueTime #54

+++ visitor-information-system/src/main/java/be/kdg/prog6/visitorInformationSystem/domain/Attraction.java
@@ -3,7 +3,7 @@ package be.kdg.prog6.visitorInformationSystem.domain;
 public class Attraction extends PointOfInterest {
     private int minHeight;
 
-    public Attraction(PointOfInterestUUID uuid, String name, String description, float posX, float posY, String picturePath, boolean open, int queueTime, int minHeight) {
+    public Attraction(PointOfInterestUUID uuid, String name, String description, float posX, float posY, String picturePath, boolean open, float queueTime, int minHeight) {
         super.setUuid(uuid);
         super.setName(name);
         super.setDescription(description);
