+++ visitor-information-system/src/main/java/be/kdg/prog6/visitorInformationSystem/domain/FoodStand.java
@@ -3,7 +3,7 @@ package be.kdg.prog6.visitorInformationSystem.domain;
 public class FoodStand extends PointOfInterest {
     private String menu;
 
-    public FoodStand(PointOfInterestUUID uuid, String name, String description, float positionX, float positionY, String picturePath, boolean open, int queueTime, String menu) {
+    public FoodStand(PointOfInterestUUID uuid, String name, String description, float positionX, float positionY, String picturePath, boolean open, float queueTime, String menu) {
         super.setUuid(uuid);
         super.setName(name);
         super.setDescription(description);
