+++ visitor-information-system/src/main/java/be/kdg/prog6/visitorInformationSystem/domain/PointOfInterest.java
@@ -11,7 +11,7 @@ public abstract class PointOfInterest {
     private float posX;
     private float posY;
     private String picturePath;
-    private int queueTime;
+    private float queueTime;
     private boolean open;
 
     public PointOfInterestUUID getUuid() {
@@ -70,11 +70,11 @@ public abstract class PointOfInterest {
         this.open = open;
     }
 
-    public int getQueueTime() {
+    public float getQueueTime() {
         return queueTime;
     }
 
-    public void setQueueTime(int queueTime) {
+    public void setQueueTime(float queueTime) {
         this.queueTime = queueTime;
     }
